# api/openapi.yaml
openapi: 3.1.0
info:
  title: uspace-ussp national USSP API
  version: 0.1.0
  description: |
    The national API of a U-space Service Provider of the Georgian
    U-space system-of-systems (Reg. (EU) 2021/664 Art. 7-13, 15). This
    file is the published contract: an endpoint that is not in it does
    not exist. The standard interfaces (ASTM F3411-22a, ASTM F3548-21)
    are served from the pinned standard files, not from this one.

    Every error is application/problem+json in the shape of the Problem
    schema (reconciliation M28).

    Skeleton (WP-0): health and readiness, the security schemes and the
    tags. Every later work package adds its paths under its tag. WP-4:
    the CIS change-notification receiver. WP-5: the cached registry
    validity lookup. WP-7: operational intents and flight authorisation.
    WP-12: geo-awareness from the CIS cache.
    WP-8: operator telemetry (served by telemetry-ingest). WP-9: the
    optional authority flight push (served by rid-sp). WP-11: traffic
    information and the alert stream (served by traffic-ws), the alert
    acknowledgement (api).
  license:
    name: Apache-2.0
servers:
  - url: "https://{host}"
    description: A USSP deployment; the host is configuration, never code.
    variables:
      host:
        default: localhost
        description: The public host of the deployment.
...
paths:
...
  /v1/cis/notifications:
    post:
      operationId: receiveCISNotification
      tags: [cis]
      summary: Receive a CIS change notification (F3 push)
      description: |
        The receiver of every subscriber in the ecosystem (M1): the CISP
        POSTs here the callback it registered, and the ANSP posts here
        on its degraded direct path while the CISP is down (M5). The body
        is a compact JWS (application/jose) whose payload is {iss, aud,
        sub, iat, jti, body}: iss one of USSP_CIS_NOTIFY_ISSUERS (each
        verified against its JWKS), aud this host (M19), sub the
        subscription (the restriction id from the ANSP), jti the
        delivery id, iat at most 5 minutes old, and body the
        cis/change/v1 record of the CISP's pinned schema. There is no
        bearer: the signature authenticates the sender.

        A delivery id seen before is acknowledged 204 and not acted on
        (it is remembered in the database on every replica). The
        reasons subscription_test and republished, and any reason this
        USSP does not know, are acknowledged 204 without a pull (M16).
        Every other reason is logged and triggers a pull of the dataset:
        from pull_url when its host is the issuer's configured base host
        (and the issuer is the CISP), otherwise from the configured CISP
        base URL, the mismatch counted (the SSRF guard). The answer does
        not wait for the pull.
      security: []
      requestBody:
        required: true
        content:
          application/jose:
            schema:
              type: string
              description: A compact JWS (three base64url parts separated by dots).
              maxLength: 262144
      responses:
        "204":
          description: Accepted, acknowledged or a replay; nothing to return.
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "413":
          $ref: "#/components/responses/Problem"
        "415":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /v1/registry/validate:
    get:
      operationId: validateRegistry
      tags: [registry]
      summary: Validity of an operator, a UAS and a remote pilot (F8, cached)
      description: |
        The authority's F8 answer through this USSP's cache, status only:
        `valid`, `suspended`, `revoked` or `unknown` per key asked, with
        the end of validity, the class label and MTOM band of a UAS and
        the competency set of a pilot, never a name, address, phone or
        e-mail (spec 02 F8, 06 §5). An answer cached within the policy's
        TTL (`registry_positive_ttl_s`, `registry_negative_ttl_s`) carries
        its `cache_age_s`; otherwise the authority is asked and the answer
        cached. When the authority cannot answer, a key without a cached
        answer is `unknown` with `reason: registry_unavailable` (or
        `registry_answer_refused` when its answer was refused), never an
        error and never `valid`. `operator` is compared on its public part
        (a secret part is neither sent, stored nor echoed); `serial` as
        given, trimmed. Every call is an `events` row
        (`registry_validated`) with the client and the purpose.
      security:
        - operatorToken: []
      parameters:
        - name: purpose
          in: query
          required: true
          schema: {$ref: "#/components/schemas/RegistryPurpose"}
        - name: operator
          in: query
          schema: {type: string, maxLength: 64}
        - name: serial
          in: query
          schema: {type: string, maxLength: 64}
        - name: pilot
          in: query
          schema: {type: string, maxLength: 32}
      responses:
        "200":
          description: One part per key asked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistryValidation"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /v1/intents:
    post:
      operationId: createIntent
      tags:
      - intents
      summary: File an operational intent and get the authorisation decision
      description: 'Decides an intent/request/v1 (2021/664 Art. 6(4), 10) in the order of brief WP-7: the
        registry (F8, purpose authorisation), the CIS cache (stale or known outdated refuses; U-space airspace
        and its Art. 3(4) constraints; zones and ANSP restrictions, Art. 10(7)), strategic deconfliction
        against the authorised intents of this USSP and the peers'' intents (priority, then first come first
        served, Art. 10(8), (9)), then the DSS, the deviation thresholds and the authorisation number (Art.
        10(11)). A missing, stale or untrusted input refuses or holds; it never authorises. An open A1 flight
        with a C0 or privately built UA below 250 g is accepted voluntarily without a number (Art. 1(3)).
        A refusal is a decision (201) whose conflicts[] name the Annex IV item, zone, airspace, restriction,
        intent or registry key; a request that does not carry the items is 400 annex_iv_invalid. Idempotent
        on (client, client_ref): the same body answers the decision as it stands (200), another body 409.
        The UAS serial must be bound to the client and operator_reg must be the client''s operator (403).
        503 names the dependency that is missing (geoid_unavailable, database_unavailable, projection_unavailable).'
      security:
      - operatorToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntentRequest'
      responses:
        '201':
          description: 'Decided: authorised, accepted_voluntary, rejected or pending_*.'
          content: &id001
            application/json:
              schema:
                $ref: '#/components/schemas/IntentDecision'
        '200':
          description: The decision of an earlier request with the same client_ref and body.
          content: *id001
        '400': &id002
          $ref: '#/components/responses/Problem'
        '401': *id002
        '403': *id002
        '409': *id002
        '413': *id002
        '429': *id002
        '503': *id002
        default: *id002
    get:
      operationId: listIntents
      tags:
      - intents
      summary: The operator's intents, newest first
      description: At most 500; from and to keep the intents whose window overlaps them; state filters.
      security:
      - operatorToken: []
      parameters:
      - name: from
        in: query
        schema:
          type: string
          format: date-time
      - name: to
        in: query
        schema:
          type: string
          format: date-time
      - name: state
        in: query
        schema:
          type: string
          maxLength: 32
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 500
      responses:
        '200':
          description: The intents.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntentList'
        '400': *id002
        '401': *id002
        '403': *id002
        '503': *id002
        default: *id002
  /v1/intents/{intent_id}:
    parameters:
    - $ref: '#/components/parameters/IntentID'
    get:
      operationId: getIntent
      tags:
      - intents
      summary: 'One of the operator''s intents: its decision and state'
      description: Another operator's intent is 404, never 403.
      security:
      - operatorToken: []
      responses:
        '200':
          description: The decision as it stands.
          content: *id001
        '401': *id002
        '403': *id002
        '404': *id002
        '503': *id002
        default: *id002
    patch:
      operationId: changeIntent
      tags:
      - intents
      summary: Activate, modify or end an intent
      description: 'activate: an accepted intent from time_start - policy.activation_lead_s to time_end,
        on the database clock, confirmed in the response (Art. 10(5)); modify: an accepted intent''s new
        volumes decided as a new version that keeps the id and, when still authorised, the authorisation
        number (Art. 6(6)); a modification that is not authorised leaves the intent in its new decision''s
        state without a number; end: any open intent. A transition the state does not allow is 409.'
      security:
      - operatorToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntentPatch'
      responses:
        '200':
          description: The new version.
          content: *id001
        '400': *id002
        '401': *id002
        '403': *id002
        '404': *id002
        '409': *id002
        '413': *id002
        '503': *id002
        default: *id002
  /v1/geo:
    get:
      operationId: getGeo
      tags: [geo]
      summary: Geo-awareness for a box at an instant
      description: |
        What the CIS says in the box (2021/664 Art. 9; spec 02 F5; brief
        WP-12), answered from this USSP's CIS cache, never from the CISP
        in the request path: the U-space airspaces with their Art. 3(4)
        requirements, the services they require and their adjacent
        airspaces; the geographical zones with their ED-318 feature
        verbatim and their limits with their references; the ANSP's
        restrictions with their state, starts_at and ends_at. Each item
        carries updated_at (its dataSource.updateDateTime, else its
        dataset version's cis_updated_at), version (the dataset version)
        and valid_from/valid_to (its limitedApplicability bounds), and how
        it applies at the instant (applies, or unknown with why: an
        unknown applicability is listed, never dropped). The answer
        carries cis_version, cis_age_s and stale (beyond the policy's
        cis_stale_s): a stale cache still answers and says so. A list cut
        at 2000 features says truncated. The geo change push of the
        traffic stream (geo/changed/v1) tells a subscriber to refetch.
      security:
      - operatorToken: []
      parameters:
      - name: bbox
        in: query
        required: true
        description: west,south,east,north in WGS84 degrees; at most 5 degrees a side.
        schema: {type: string, maxLength: 128}
      - name: at
        in: query
        required: false
        description: The instant (RFC 3339); now when absent.
        schema: {type: string, format: date-time}
      responses:
        '200':
          description: The answer from the CIS cache.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeoAnswer'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/geo/intents/{intent_id}:
    parameters:
    - $ref: '#/components/parameters/IntentID'
    get:
      operationId: getGeoForIntent
      tags: [geo]
      summary: Geo-awareness for one of the operator's intents
      description: |
        GET /v1/geo with the intent's volumes and their windows as the
        query: every feature whose extent meets a volume's box and that
        applies at some time of its window (its applicability says how:
        always, during from/to, or scheduled on daily windows from/to). The
        operator's own intent only: another operator's is 404, never 403.
      security:
      - operatorToken: []
      responses:
        '200':
          description: The answer from the CIS cache.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeoAnswer'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
...
  /v1/alerts/{alert_id}/ack:
    post:
      operationId: ackAlert
      tags: [alerts]
      summary: Acknowledge an alert
      description: |
        Records, on the database clock, that the operator acknowledged an
        alert of one of its flights (served by api; Art. 13, 02 F5): the
        first acknowledgement is kept and a repeat answers it again. The
        acknowledged alert is republished to traffic-ws, which stops
        repeating it, and a critical alert acknowledged within
        escalation_after_s (30 s) is not escalated. An alert of another
        operator's flight is 404, never 403. No body.
      security:
      - operatorToken: []
      parameters:
      - $ref: "#/components/parameters/AlertID"
      responses:
        "200":
          description: Acknowledged.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertAck"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /oauth/token:
    post:
      operationId: requestToken
      tags: [oauth]
      summary: Client credentials grant for operator machine clients
      description: |
        RFC 6749 §4.4 client credentials at this USSP's own issuer (D9).
        The client authenticates with HTTP Basic (`client_secret_basic`)
        or with `client_id` and `client_secret` in the body
        (`client_secret_post`), never both; the secret is checked with
        argon2id in constant time, and during a rotation the previous
        secret is accepted until its overlap ends
        (`policy.client_secret_overlap_s`). Parameters in the query
        string are refused. `scope` names operator scopes (`ussp.intents`,
        `ussp.telemetry`, `ussp.traffic`, `ussp.geo`); an unknown scope is
        refused; the token grants the requested scopes that the client
        holds (all of them when `scope` is absent). `audience`, when
        given, must be this USSP's host: `aud` is always this USSP's host
        (M18). The TTL is `policy.operator_token_ttl_s` (at most 3600 s).
        Issuance is rate-limited per client and per client address (429
        with `Retry-After`). Every issuance and refusal is an `events`
        row. Errors carry the RFC 6749 §5.2 `error` beside the problem
        members.
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/TokenRequest"
      responses:
        "200":
          description: An operator machine token (cross-plan Appendix A).
          headers:
            Cache-Control:
              schema: {type: string, enum: [no-store]}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          $ref: "#/components/responses/OAuthError"
        "401":
          $ref: "#/components/responses/OAuthError"
        "429":
          description: The client's or the address's issuance budget is spent.
          headers:
            Retry-After:
              schema: {type: integer, minimum: 1}
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/OAuthProblem"
        "503":
          $ref: "#/components/responses/OAuthError"
        default:
          $ref: "#/components/responses/Problem"
  /.well-known/jwks.json:
    get:
      operationId: getJWKS
      tags: [oauth]
      summary: The issuer's public keys
      description: |
        The RS256 keys of this USSP's issuer (`use: sig`, `kid` = the RFC
        7638 SHA-256 thumbprint): the current key and, during a rotation,
        the previous one (`USSP_ISSUER_PREVIOUS_KEY_FILE`). Operator
        tokens and session tokens are signed by the current key.
      security: []
      responses:
        "200":
          description: The JWKS.
          headers:
            Cache-Control:
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JWKS"
        "503":
          description: No issuer key is configured.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    operatorToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        RS256 machine token issued by this USSP (POST /oauth/token,
        client credentials) to an operator client; aud is this host.
    ecosystemToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        RS256 machine token of an allow-listed ecosystem issuer (the
        authority's token service; the lab issuer in the lab) presented
        by the authority, the ANSP or a peer; aud is this host (M18).
    session:
      type: apiKey
      in: cookie
      name: uspace_session
      description: |
        Portal or console session: a JWT with scope "session", roles and
        realm, in the uspace_session cookie; state-changing requests also
        send the uspace_csrf cookie's value as X-CSRF-Token (M20, M21).
...